Class files moved to views/entities #258
