+++ application/android/app/src/main/java/com/group12/beabee/views/entities/QuestionFragment.java
@@ -1,4 +1,4 @@
-package com.group12.beabee;
+package com.group12.beabee.views.entities;
 
 import android.os.Bundle;
 
@@ -7,20 +7,15 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 
+import com.group12.beabee.R;
 import com.group12.beabee.models.QuestionShort;
 import com.group12.beabee.models.ReflectionShort;
 import com.group12.beabee.models.RoutineShort;
 import com.group12.beabee.models.TaskShort;
 import com.group12.beabee.views.BaseInnerFragment;
 import com.group12.beabee.views.MainPage.PageMode;
-import com.group12.beabee.views.entities.IOnQuestionClickedListener;
-import com.group12.beabee.views.entities.IOnReflectionClickedListener;
-import com.group12.beabee.views.entities.IOnRoutineClickedListener;
-import com.group12.beabee.views.entities.IOnTaskClickedListener;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -29,10 +24,10 @@ import butterknife.BindView;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link RoutineFragment#newInstance} factory method to
+ * Use the {@link QuestionFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class RoutineFragment extends BaseInnerFragment implements IOnTaskClickedListener,
+public class QuestionFragment extends BaseInnerFragment implements IOnTaskClickedListener,
         IOnRoutineClickedListener, IOnQuestionClickedListener, IOnReflectionClickedListener {
 
 
@@ -45,7 +40,7 @@ public class RoutineFragment extends BaseInnerFragment implements IOnTaskClicked
     @BindView(R.id.rv_reflections)
     RecyclerView rvReflections;
 
-    public RoutineFragment() {
+    public QuestionFragment() {
         // Required empty public constructor
     }
 
@@ -53,22 +48,20 @@ public class RoutineFragment extends BaseInnerFragment implements IOnTaskClicked
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
-     * @return A new instance of fragment Routine.
+     * @return A new instance of fragment Question.
      */
     // TODO: Rename and change types and number of parameters
-    public static RoutineFragment newInstance() {
-        RoutineFragment fragment = new RoutineFragment();
+    public static QuestionFragment newInstance() {
+        QuestionFragment fragment = new QuestionFragment();
         Bundle args = new Bundle();
         fragment.setArguments(args);
         return fragment;
     }
 
 
-
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-
 
         TaskCardViewAdapter tasksAdapter = new TaskCardViewAdapter();
         RoutineCardViewAdapter routinesAdapter = new RoutineCardViewAdapter();
@@ -171,6 +164,7 @@ public class RoutineFragment extends BaseInnerFragment implements IOnTaskClicked
 
     @Override
     protected int GetLayoutId() {
-        return R.layout.fragment_routine;
+        return R.layout.fragment_question;
     }
+
 }
