+++ application/android/app/src/main/java/com/group12/beabee/views/entities/QuestionFragmentEdit.java
@@ -1,25 +1,23 @@
-package com.group12.beabee;
+package com.group12.beabee.views.entities;
 
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.RecyclerView;
 
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 
+import com.group12.beabee.R;
 import com.group12.beabee.views.BaseInnerFragment;
 import com.group12.beabee.views.MainPage.PageMode;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link ReflectionFragmentEdit#newInstance} factory method to
+ * Use the {@link QuestionFragmentEdit#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class ReflectionFragmentEdit extends BaseInnerFragment {
+public class QuestionFragmentEdit extends BaseInnerFragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -31,7 +29,7 @@ public class ReflectionFragmentEdit extends BaseInnerFragment {
     private String mParam2;
 
 
-    public ReflectionFragmentEdit() {
+    public QuestionFragmentEdit() {
         // Required empty public constructor
     }
 
@@ -39,13 +37,11 @@ public class ReflectionFragmentEdit extends BaseInnerFragment {
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment ReflectionEdit.
+     * @return A new instance of fragment QuestionEdit.
      */
     // TODO: Rename and change types and number of parameters
-    public static ReflectionFragmentEdit newInstance() {
-        ReflectionFragmentEdit fragment = new ReflectionFragmentEdit();
+    public static QuestionFragmentEdit newInstance() {
+        QuestionFragmentEdit fragment = new QuestionFragmentEdit();
         Bundle args = new Bundle();
         fragment.setArguments(args);
         return fragment;
@@ -66,6 +62,6 @@ public class ReflectionFragmentEdit extends BaseInnerFragment {
 
     @Override
     protected int GetLayoutId() {
-        return R.layout.fragment_reflection_edit;
+        return R.layout.fragment_question_edit;
     }
 }
