+++ application/android/app/src/main/java/com/group12/beabee/views/entities/ReflectionFragmentEdit.java
@@ -1,25 +1,23 @@
-package com.group12.beabee;
+package com.group12.beabee.views.entities;
 
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
-import androidx.recyclerview.widget.RecyclerView;
 
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 
+import com.group12.beabee.R;
 import com.group12.beabee.views.BaseInnerFragment;
 import com.group12.beabee.views.MainPage.PageMode;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link TaskFragmentEdit#newInstance} factory method to
+ * Use the {@link ReflectionFragmentEdit#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class TaskFragmentEdit extends BaseInnerFragment {
+public class ReflectionFragmentEdit extends BaseInnerFragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -31,7 +29,7 @@ public class TaskFragmentEdit extends BaseInnerFragment {
     private String mParam2;
 
 
-    public TaskFragmentEdit() {
+    public ReflectionFragmentEdit() {
         // Required empty public constructor
     }
 
@@ -41,11 +39,11 @@ public class TaskFragmentEdit extends BaseInnerFragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment TaskEdit.
+     * @return A new instance of fragment ReflectionEdit.
      */
     // TODO: Rename and change types and number of parameters
-    public static TaskFragmentEdit newInstance() {
-        TaskFragmentEdit fragment = new TaskFragmentEdit();
+    public static ReflectionFragmentEdit newInstance() {
+        ReflectionFragmentEdit fragment = new ReflectionFragmentEdit();
         Bundle args = new Bundle();
         fragment.setArguments(args);
         return fragment;
@@ -66,6 +64,6 @@ public class TaskFragmentEdit extends BaseInnerFragment {
 
     @Override
     protected int GetLayoutId() {
-        return R.layout.fragment_task_edit;
+        return R.layout.fragment_reflection_edit;
     }
 }
