+++ application/android/app/src/main/java/com/group12/beabee/views/entities/RoutineCardViewAdapter.java
@@ -1,4 +1,4 @@
-package com.group12.beabee;
+package com.group12.beabee.views.entities;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -8,9 +8,8 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.group12.beabee.R;
 import com.group12.beabee.models.RoutineShort;
-
-import com.group12.beabee.views.entities.IOnRoutineClickedListener;
 
 import java.util.List;
