+++ application/android/app/src/main/java/com/group12/beabee/views/entities/TaskFragment.java
@@ -1,4 +1,4 @@
-package com.group12.beabee;
+package com.group12.beabee.views.entities;
 
 import android.os.Bundle;
 
@@ -7,20 +7,15 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 
+import com.group12.beabee.R;
 import com.group12.beabee.models.QuestionShort;
 import com.group12.beabee.models.ReflectionShort;
 import com.group12.beabee.models.RoutineShort;
 import com.group12.beabee.models.TaskShort;
 import com.group12.beabee.views.BaseInnerFragment;
 import com.group12.beabee.views.MainPage.PageMode;
-import com.group12.beabee.views.entities.IOnQuestionClickedListener;
-import com.group12.beabee.views.entities.IOnReflectionClickedListener;
-import com.group12.beabee.views.entities.IOnRoutineClickedListener;
-import com.group12.beabee.views.entities.IOnTaskClickedListener;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -29,11 +24,12 @@ import butterknife.BindView;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link ReflectionFragment#newInstance} factory method to
+ * Use the {@link TaskFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class ReflectionFragment extends BaseInnerFragment implements IOnTaskClickedListener,
+public class TaskFragment extends BaseInnerFragment implements IOnTaskClickedListener,
         IOnRoutineClickedListener, IOnQuestionClickedListener, IOnReflectionClickedListener {
+
 
     @BindView(R.id.rv_tasks)
     RecyclerView rvTasks;
@@ -44,7 +40,7 @@ public class ReflectionFragment extends BaseInnerFragment implements IOnTaskClic
     @BindView(R.id.rv_reflections)
     RecyclerView rvReflections;
 
-    public ReflectionFragment() {
+    public TaskFragment() {
         // Required empty public constructor
     }
 
@@ -52,15 +48,18 @@ public class ReflectionFragment extends BaseInnerFragment implements IOnTaskClic
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
-     * @return A new instance of fragment Reflection.
+     * @return A new instance of fragment Task.
      */
     // TODO: Rename and change types and number of parameters
-    public static ReflectionFragment newInstance() {
-        ReflectionFragment fragment = new ReflectionFragment();
+    public static TaskFragment newInstance() {
+        TaskFragment fragment = new TaskFragment();
         Bundle args = new Bundle();
         fragment.setArguments(args);
         return fragment;
     }
+
+
+
 
 
     @Override
@@ -161,6 +160,7 @@ public class ReflectionFragment extends BaseInnerFragment implements IOnTaskClic
         //sendreqquest for reflectiondata Open
     }
 
+
     @Override
     protected PageMode GetPageMode() {
         return PageMode.Editable;
@@ -168,6 +168,6 @@ public class ReflectionFragment extends BaseInnerFragment implements IOnTaskClic
 
     @Override
     protected int GetLayoutId() {
-        return R.layout.fragment_reflection;
+        return R.layout.fragment_task;
     }
 }
